+++ mello/mello/App.js
@@ -6,6 +6,7 @@ import { NavigationContainer } from '@react-navigation/native';
 import Home from './Home'
 import Settings from './Settings';
 import TimeManagement from './TimeManagement';
+import WeekView from './WeekView';
 
 
 const Drawer = createDrawerNavigator();
@@ -23,6 +24,7 @@ export default function App() {
            } }} />
       <Drawer.Screen name="Settings" component={Settings} />
       <Drawer.Screen name="TimeManagement" component={TimeManagement} options={{ headerShown: false}} />
+      <Drawer.Screen name="Daily Planner" component={WeekView} options={{ headerShown: false}} />
     </Drawer.Navigator>
   </NavigationContainer>
   );
